+++ mooda/__init__.py
@@ -2,7 +2,7 @@ from mooda.waterframe import WaterFrame
 try:
     from mooda.plotmap import PlotMap
 except ImportError:
-    print("Warning: You do not have installed the library basemap. You cannot",
+    print("Warning: You do not have the basemap library installed. You cannot",
           "use PlotMap.")
 
 name = "mooda"